Refactor transcript, fsm, handshake to internal
