+++ internal/handshake/fsm_test.go
@@ -1,13 +1,14 @@
 // SPDX-FileCopyrightText: 2026 The Pion community <https://pion.ly>
 // SPDX-License-Identifier: MIT
 
-package dtls
+package dtlshandshake
 
 import (
 	"context"
 	"testing"
 	"time"
 
+	"github.com/pion/dtls/v3/internal/ciphersuite"
 	dtlsconfig "github.com/pion/dtls/v3/internal/config"
 	dtlserrors "github.com/pion/dtls/v3/internal/errors"
 	dtlsflight "github.com/pion/dtls/v3/internal/flight"
@@ -16,6 +17,7 @@ import (
 	"github.com/pion/dtls/v3/pkg/crypto/elliptic"
 	"github.com/pion/dtls/v3/pkg/crypto/signaturehash"
 	"github.com/pion/dtls/v3/pkg/protocol"
+	"github.com/pion/dtls/v3/pkg/protocol/alert"
 	"github.com/pion/dtls/v3/pkg/protocol/extension"
 	"github.com/pion/dtls/v3/pkg/protocol/handshake"
 	"github.com/pion/dtls/v3/pkg/protocol/recordlayer"
@@ -24,12 +26,69 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+var testCurves13 = []elliptic.Curve{elliptic.X25519, elliptic.P256, elliptic.P384} //nolint:gochecknoglobals
+
+type handshakeContext13 struct {
+	state      *dtlsstate.State
+	cache      *dtlsflight.Cache
+	cfg        *dtlsconfig.HandshakeConfig
+	transcript *Transcript
+}
+
+func (s *fsm13) flightContext() *handshakeContext13 {
+	return &handshakeContext13{
+		state:      s.state,
+		cache:      s.cache,
+		cfg:        s.cfg,
+		transcript: s.transcript,
+	}
+}
+
+func flight13GenerateForTest(
+	testingT require.TestingT,
+	flight dtlsflight13.Flight,
+	flightCtx *handshakeContext13,
+) ([]*dtlsflight.Packet, *alert.Alert, error) {
+	if helper, ok := testingT.(interface{ Helper() }); ok {
+		helper.Helper()
+	}
+
+	gen, _, ok := dtlsflight13.GetGenerator(flight)
+	require.True(testingT, ok)
+
+	return gen(nil, flightCtx.state, flightCtx.cache, flightCtx.cfg)
+}
+
+type flightTestConn struct{}
+
+func (c *flightTestConn) Notify(context.Context, alert.Level, alert.Description) error {
+	return nil
+}
+
+func (c *flightTestConn) WritePackets(context.Context, []*dtlsflight.Packet) error {
+	return nil
+}
+
+func (c *flightTestConn) RecvHandshake() <-chan RecvHandshakeState {
+	return nil
+}
+
+func (c *flightTestConn) SetLocalEpoch(uint16) {}
+
+func (c *flightTestConn) HandleQueuedPackets(context.Context) error {
+	return nil
+}
+
+func (c *flightTestConn) SessionKey() []byte {
+	return nil
+}
+
 func TestHandshakeFSM13OwnsTranscriptAndPropagatesContext(t *testing.T) {
 	state := &dtlsstate.State{IsClient: true, LocalVersion: protocol.Version1_3}
 	cache := dtlsflight.NewCache()
 	cfg := testHandshakeConfig13(t)
 
-	fsm, err := newHandshakeFSM13(state, cache, cfg, dtlsflight13.Flight1, nil, nil)
+	fsm, err := newFSM13(state, cache, cfg, dtlsflight13.Flight1, nil, nil)
 	require.NoError(t, err)
 	require.NotNil(t, fsm.transcript)
 
@@ -50,7 +109,7 @@ func TestHandshakeFSM13DualStackClientHelloSeedsTranscript(t *testing.T) {
 		return &ch
 	}
 
-	transcript := newHandshakeTranscript13()
+	transcript := NewTranscript()
 	pkts, dtlsAlert, err := flight13GenerateForTest(t, dtlsflight13.Flight1, &handshakeContext13{
 		state:      state,
 		cache:      cache,
@@ -67,31 +126,31 @@ func TestHandshakeFSM13DualStackClientHelloSeedsTranscript(t *testing.T) {
 	content.Header.MessageSequence = messageSequence
 
 	expected := canonicalPacketHandshake13(t, pkts[0])
-	appended, err := appendClientHelloInitialFlights13(transcript, pkts)
+	appended, err := AppendClientHelloInitialFlights(transcript, pkts)
 	require.NoError(t, err)
 	require.True(t, appended)
 
-	fsm, err := newHandshakeFSM13(state, cache, cfg, dtlsflight13.Flight1, pkts, transcript)
+	fsm, err := newFSM13(state, cache, cfg, dtlsflight13.Flight1, pkts, transcript)
 	require.NoError(t, err)
 	require.NotNil(t, fsm.transcript)
 	require.Same(t, transcript, fsm.transcript)
-	require.Len(t, fsm.transcript.pending, 1)
-	require.Len(t, fsm.transcript.order, 1)
+	require.Len(t, fsm.transcript.pendingMessages(), 1)
+	require.Len(t, fsm.transcript.messageOrder(), 1)
 
-	assert.Equal(t, expected, fsm.transcript.pending[0])
-	assert.Equal(t, expected, fsm.transcript.transcript)
-	assert.Equal(t, transcriptMessageID13{
-		sender: transcriptClient13,
-		seq:    messageSequence,
-	}, fsm.transcript.order[0].id)
-	assert.Equal(t, handshake.TypeClientHello, fsm.transcript.order[0].typ)
+	assert.Equal(t, expected, fsm.transcript.pendingMessages()[0])
+	assert.Equal(t, expected, fsm.transcript.Bytes())
+	assert.Equal(t, transcriptMessageID{
+		sender: transcriptSenderClient,
+		Seq:    messageSequence,
+	}, fsm.transcript.messageOrder()[0].ID)
+	assert.Equal(t, handshake.TypeClientHello, fsm.transcript.messageOrder()[0].Type)
 }
 
 func TestHandshakeFSM13TranscriptSurvivesStateChangesAndRetransmitSeed(t *testing.T) {
 	state := &dtlsstate.State{IsClient: true, LocalVersion: protocol.Version1_3}
 	cache := dtlsflight.NewCache()
 	cfg := testHandshakeConfig13(t)
-	transcript := newHandshakeTranscript13()
+	transcript := NewTranscript()
 
 	pkts, dtlsAlert, err := flight13GenerateForTest(t, dtlsflight13.Flight1, &handshakeContext13{
 		state:      state,
@@ -102,25 +161,25 @@ func TestHandshakeFSM13TranscriptSurvivesStateChangesAndRetransmitSeed(t *testin
 	require.NoError(t, err)
 	require.Nil(t, dtlsAlert)
 
-	fsm, err := newHandshakeFSM13(state, cache, cfg, dtlsflight13.Flight1, pkts, transcript)
+	fsm, err := newFSM13(state, cache, cfg, dtlsflight13.Flight1, pkts, transcript)
 	require.NoError(t, err)
 
 	transcript = fsm.transcript
-	before := append([]byte(nil), transcript.transcript...)
-	require.Len(t, transcript.pending, 1)
+	before := append([]byte(nil), transcript.Bytes()...)
+	require.Len(t, transcript.pendingMessages(), 1)
 
 	fsm.currentFlight = dtlsflight13.Flight2
 	fsm.retransmit = true
 	fsm.retransmitInterval *= 2
 
 	assert.Same(t, transcript, fsm.transcript)
-	assert.Equal(t, before, fsm.transcript.transcript)
+	assert.Equal(t, before, fsm.transcript.Bytes())
 	assert.Same(t, transcript, fsm.flightContext().transcript)
 
 	require.NoError(t, fsm.seedTranscriptFromInitialFlights())
 	assert.Same(t, transcript, fsm.transcript)
-	assert.Equal(t, before, fsm.transcript.transcript)
-	assert.Len(t, fsm.transcript.pending, 1)
+	assert.Equal(t, before, fsm.transcript.Bytes())
+	assert.Len(t, fsm.transcript.pendingMessages(), 1)
 }
 
 func TestHandshakeFSM13DualStackClientHelloRequired(t *testing.T) {
@@ -128,8 +187,8 @@ func TestHandshakeFSM13DualStackClientHelloRequired(t *testing.T) {
 	cache := dtlsflight.NewCache()
 	cfg := testHandshakeConfig13(t)
 
-	fsm, err := newHandshakeFSM13(
-		state, cache, cfg, dtlsflight13.Flight1, []*dtlsflight.Packet{}, newHandshakeTranscript13(),
+	fsm, err := newFSM13(
+		state, cache, cfg, dtlsflight13.Flight1, []*dtlsflight.Packet{}, NewTranscript(),
 	)
 	require.Nil(t, fsm)
 	require.ErrorIs(t, err, dtlserrors.ErrHandshakeTranscriptMissingClientHello)
@@ -143,15 +202,15 @@ func TestHandshakeFSM13PrepareHelloRetryRequestRequiresSeededTranscript(t *testi
 	}
 	cache := dtlsflight.NewCache()
 
-	fsm, err := newHandshakeFSM13(state, cache, cfg, dtlsflight13.Flight2, nil, nil)
+	fsm, err := newFSM13(state, cache, cfg, dtlsflight13.Flight2, nil, nil)
 	require.NoError(t, err)
 
 	nextState, err := fsm.prepare(context.Background(), nil)
 	require.ErrorIs(t, err, dtlserrors.ErrHandshakeTranscriptHelloRetryRequestInvalid)
-	assert.Equal(t, handshakeErrored, nextState)
+	assert.Equal(t, StateErrored, nextState)
 	require.Len(t, fsm.flights, 1)
-	assert.Empty(t, fsm.transcript.order)
-	assert.Empty(t, fsm.transcript.transcript)
+	assert.Empty(t, fsm.transcript.messageOrder())
+	assert.Empty(t, fsm.transcript.Bytes())
 	assert.Equal(t, 1, state.HandshakeSendSequence)
 }
 
@@ -160,19 +219,19 @@ func TestHandshakeFSM13PrepareCommitsOutboundClientHello(t *testing.T) {
 	state := &dtlsstate.State{IsClient: true, LocalVersion: protocol.Version1_3}
 	cache := dtlsflight.NewCache()
 
-	fsm, err := newHandshakeFSM13(state, cache, cfg, dtlsflight13.Flight1, nil, nil)
+	fsm, err := newFSM13(state, cache, cfg, dtlsflight13.Flight1, nil, nil)
 	require.NoError(t, err)
 
 	nextState, err := fsm.prepare(context.Background(), nil)
 	require.NoError(t, err)
-	assert.Equal(t, handshakeSending, nextState)
+	assert.Equal(t, StateSending, nextState)
 	require.Len(t, fsm.flights, 1)
 
 	expected := canonicalPacketHandshake13(t, fsm.flights[0])
-	require.Len(t, fsm.transcript.order, 1)
-	assert.Equal(t, transcriptMessageID13{sender: transcriptClient13, seq: 0}, fsm.transcript.order[0].id)
-	assert.Equal(t, handshake.TypeClientHello, fsm.transcript.order[0].typ)
-	assert.Equal(t, expected, fsm.transcript.transcript)
+	require.Len(t, fsm.transcript.messageOrder(), 1)
+	assert.Equal(t, transcriptMessageID{sender: transcriptSenderClient, Seq: 0}, fsm.transcript.messageOrder()[0].ID)
+	assert.Equal(t, handshake.TypeClientHello, fsm.transcript.messageOrder()[0].Type)
+	assert.Equal(t, expected, fsm.transcript.Bytes())
 	assert.Equal(t, 1, state.HandshakeSendSequence)
 }
 
@@ -183,27 +242,27 @@ func TestHandshakeFSM13PrepareCommitsOutboundHelloRetryRequestWithSeededTranscri
 		CipherSuite:  cfg.LocalCipherSuites[0],
 	}
 	cache := dtlsflight.NewCache()
-	transcript := newHandshakeTranscript13()
+	transcript := NewTranscript()
 	clientHello := transcriptTestClientHelloPacket13([]byte{0x01}, 0)
 	clientHelloCanonical := canonicalPacketHandshake13(t, clientHello)
-	require.NoError(t, appendOutboundHandshakeFlight13(transcript, true, nil, []*dtlsflight.Packet{clientHello}))
+	require.NoError(t, AppendOutboundHandshakeFlight(transcript, true, nil, []*dtlsflight.Packet{clientHello}))
 
-	fsm, err := newHandshakeFSM13(state, cache, cfg, dtlsflight13.Flight2, nil, transcript)
+	fsm, err := newFSM13(state, cache, cfg, dtlsflight13.Flight2, nil, transcript)
 	require.NoError(t, err)
 
 	nextState, err := fsm.prepare(context.Background(), nil)
 	require.NoError(t, err)
-	assert.Equal(t, handshakeSending, nextState)
+	assert.Equal(t, StateSending, nextState)
 	require.Len(t, fsm.flights, 1)
 
 	helloRetryRequestCanonical := canonicalPacketHandshake13(t, fsm.flights[0])
 	messageHash := canonicalTranscriptHandshake13(handshake.TypeMessageHash, hashTranscript13(clientHelloCanonical))
 	expectedTranscript := append(append([]byte(nil), messageHash...), helloRetryRequestCanonical...)
 
-	assert.Equal(t, expectedTranscript, fsm.transcript.transcript)
-	require.Len(t, fsm.transcript.order, 2)
-	assert.Equal(t, transcriptMessageID13{sender: transcriptServer13, seq: 0}, fsm.transcript.order[1].id)
-	assert.Equal(t, handshake.TypeServerHello, fsm.transcript.order[1].typ)
+	assert.Equal(t, expectedTranscript, fsm.transcript.Bytes())
+	require.Len(t, fsm.transcript.messageOrder(), 2)
+	assert.Equal(t, transcriptMessageID{sender: transcriptSenderServer, Seq: 0}, fsm.transcript.messageOrder()[1].ID)
+	assert.Equal(t, handshake.TypeServerHello, fsm.transcript.messageOrder()[1].Type)
 	assert.Equal(t, 1, state.HandshakeSendSequence)
 }
 
@@ -220,31 +279,31 @@ func TestHandshakeFSM13PrepareDerivesTrafficSecretsBeforeEncryptedExtensions(t *
 		LocalRandom:     handshake.Random{RandomBytes: [handshake.RandomBytesLength]byte{0x01}},
 		PreMasterSecret: []byte{0x01, 0x02, 0x03},
 	}
-	transcript := newHandshakeTranscript13()
+	transcript := NewTranscript()
 	clientHello := transcriptTestClientHelloPacket13([]byte{0x01}, 0)
 	clientHelloCanonical := canonicalPacketHandshake13(t, clientHello)
-	require.NoError(t, appendOutboundHandshakeFlight13(transcript, true, nil, []*dtlsflight.Packet{clientHello}))
+	require.NoError(t, AppendOutboundHandshakeFlight(transcript, true, nil, []*dtlsflight.Packet{clientHello}))
 
-	fsm, err := newHandshakeFSM13(state, dtlsflight.NewCache(), cfg, dtlsflight13.Flight4, nil, transcript)
+	fsm, err := newFSM13(state, dtlsflight.NewCache(), cfg, dtlsflight13.Flight4, nil, transcript)
 	require.NoError(t, err)
 
 	nextState, err := fsm.prepare(context.Background(), &flightTestConn{})
 	require.NoError(t, err)
-	assert.Equal(t, handshakeSending, nextState)
+	assert.Equal(t, StateSending, nextState)
 	require.Len(t, fsm.flights, 2)
 
 	serverHelloCanonical := canonicalPacketHandshake13(t, fsm.flights[0])
 	encryptedExtensionsCanonical := canonicalPacketHandshake13(t, fsm.flights[1])
 	expectedTranscript := append(append(append([]byte(nil), clientHelloCanonical...), serverHelloCanonical...),
 		encryptedExtensionsCanonical...)
-	assert.Equal(t, expectedTranscript, fsm.transcript.transcript)
-	assert.Equal(t, []transcriptMessage13{
-		{id: transcriptMessageID13{sender: transcriptClient13, seq: 0}, typ: handshake.TypeClientHello},
-		{id: transcriptMessageID13{sender: transcriptServer13, seq: 0}, typ: handshake.TypeServerHello},
-		{id: transcriptMessageID13{sender: transcriptServer13, seq: 1}, typ: handshake.TypeEncryptedExtensions},
-	}, fsm.transcript.order)
+	assert.Equal(t, expectedTranscript, fsm.transcript.Bytes())
+	assert.Equal(t, []transcriptMessage{
+		{ID: transcriptMessageID{sender: transcriptSenderClient, Seq: 0}, Type: handshake.TypeClientHello},
+		{ID: transcriptMessageID{sender: transcriptSenderServer, Seq: 0}, Type: handshake.TypeServerHello},
+		{ID: transcriptMessageID{sender: transcriptSenderServer, Seq: 1}, Type: handshake.TypeEncryptedExtensions},
+	}, fsm.transcript.messageOrder())
 
-	expectedSecrets, err := deriveHandshakeTrafficSecrets13(
+	expectedSecrets, err := deriveHandshakeTrafficSecrets(
 		state.CipherSuite.HashFunc(),
 		state.PreMasterSecret,
 		hashTranscript13(clientHelloCanonical, serverHelloCanonical),
@@ -260,32 +319,25 @@ func canonicalPacketHandshake13(t *testing.T, p *dtlsflight.Packet) []byte {
 	require.True(t, ok)
 	raw, err := content.Marshal()
 	require.NoError(t, err)
-	canonical, err := canonicalHandshake13(raw)
+	canonical, err := canonicalHandshake(raw)
 	require.NoError(t, err)
 
 	return canonical
 }
 
-func testHandshakeConfig13(t *testing.T) *handshakeConfig {
+func testHandshakeConfig13(t *testing.T) *dtlsconfig.HandshakeConfig {
 	t.Helper()
 
-	cipherSuites, err := parseCipherSuitesForVersions(
-		nil,
-		nil,
-		true,
-		false,
-		protocol.Version1_3,
-		protocol.Version1_3,
-	)
-	require.NoError(t, err)
+	cipherSuite := ciphersuite.ForID(ciphersuite.TLS_AES_128_GCM_SHA256, nil)
+	require.NotNil(t, cipherSuite)
 
 	loggerFactory := logging.NewDefaultLoggerFactory()
 
-	return &handshakeConfig{
-		LocalCipherSuites:           cipherSuites,
-		EllipticCurves:              defaultCurves,
+	return &dtlsconfig.HandshakeConfig{
+		LocalCipherSuites:           []dtlsconfig.CipherSuite{cipherSuite},
+		EllipticCurves:              testCurves13,
 		InitialRetransmitInterval:   time.Second,
-		ExtendedMasterSecret:        dtlsconfig.ExtendedMasterSecretType(RequestExtendedMasterSecret),
+		ExtendedMasterSecret:        dtlsconfig.RequestExtendedMasterSecret,
 		Log:                         loggerFactory.NewLogger("dtls"),
 		MinVersion:                  protocol.Version1_3,
 		MaxVersion:                  protocol.Version1_3,
@@ -296,40 +348,40 @@ func testHandshakeConfig13(t *testing.T) *handshakeConfig {
 }
 
 func TestAppendOutboundHandshakeFlight13ClientHello(t *testing.T) {
-	transcript := newHandshakeTranscript13()
+	transcript := NewTranscript()
 	pkt := transcriptTestClientHelloPacket13([]byte{0x01}, 3)
 	expected := canonicalPacketHandshake13(t, pkt)
 
-	err := appendOutboundHandshakeFlight13(transcript, true, nil, []*dtlsflight.Packet{pkt})
+	err := AppendOutboundHandshakeFlight(transcript, true, nil, []*dtlsflight.Packet{pkt})
 	require.NoError(t, err)
-	require.Len(t, transcript.order, 1)
-	require.Len(t, transcript.pending, 1)
+	require.Len(t, transcript.messageOrder(), 1)
+	require.Len(t, transcript.pendingMessages(), 1)
 
-	assert.Equal(t, transcriptMessageID13{sender: transcriptClient13, seq: 3}, transcript.order[0].id)
-	assert.Equal(t, handshake.TypeClientHello, transcript.order[0].typ)
-	assert.Equal(t, expected, transcript.pending[0])
-	assert.Equal(t, expected, transcript.transcript)
+	assert.Equal(t, transcriptMessageID{sender: transcriptSenderClient, Seq: 3}, transcript.messageOrder()[0].ID)
+	assert.Equal(t, handshake.TypeClientHello, transcript.messageOrder()[0].Type)
+	assert.Equal(t, expected, transcript.pendingMessages()[0])
+	assert.Equal(t, expected, transcript.Bytes())
 }
 
 func TestAppendOutboundHandshakeFlight13DuplicateNoop(t *testing.T) {
-	transcript := newHandshakeTranscript13()
+	transcript := NewTranscript()
 	pkt := transcriptTestClientHelloPacket13([]byte{0x01}, 0)
 
-	require.NoError(t, appendOutboundHandshakeFlight13(transcript, true, nil, []*dtlsflight.Packet{pkt}))
-	before := append([]byte(nil), transcript.transcript...)
+	require.NoError(t, AppendOutboundHandshakeFlight(transcript, true, nil, []*dtlsflight.Packet{pkt}))
+	before := append([]byte(nil), transcript.Bytes()...)
 
-	require.NoError(t, appendOutboundHandshakeFlight13(transcript, true, nil, []*dtlsflight.Packet{pkt}))
-	assert.Equal(t, before, transcript.transcript)
-	assert.Len(t, transcript.order, 1)
+	require.NoError(t, AppendOutboundHandshakeFlight(transcript, true, nil, []*dtlsflight.Packet{pkt}))
+	assert.Equal(t, before, transcript.Bytes())
+	assert.Len(t, transcript.messageOrder(), 1)
 }
 
 func TestAppendOutboundHandshakeFlight13ChangedSameSequenceFails(t *testing.T) {
-	transcript := newHandshakeTranscript13()
+	transcript := NewTranscript()
 	pkt := transcriptTestClientHelloPacket13([]byte{0x01}, 0)
 	changedPkt := transcriptTestClientHelloPacket13([]byte{0x02}, 0)
 
-	require.NoError(t, appendOutboundHandshakeFlight13(transcript, true, nil, []*dtlsflight.Packet{pkt}))
-	err := appendOutboundHandshakeFlight13(transcript, true, nil, []*dtlsflight.Packet{changedPkt})
+	require.NoError(t, AppendOutboundHandshakeFlight(transcript, true, nil, []*dtlsflight.Packet{pkt}))
+	err := AppendOutboundHandshakeFlight(transcript, true, nil, []*dtlsflight.Packet{changedPkt})
 
 	assert.ErrorIs(t, err, dtlserrors.ErrHandshakeTranscriptMessageChanged)
 }
@@ -337,36 +389,36 @@ func TestAppendOutboundHandshakeFlight13ChangedSameSequenceFails(t *testing.T) {
 func TestAppendOutboundHandshakeFlight13HelloRetryRequest(t *testing.T) {
 	cfg := testHandshakeConfig13(t)
 	cipherSuite := cfg.LocalCipherSuites[0]
-	transcript := newHandshakeTranscript13()
+	transcript := NewTranscript()
 	clientHello := transcriptTestClientHelloPacket13([]byte{0x01}, 0)
 	helloRetryRequest := transcriptTestHelloRetryRequestPacket13(t, cipherSuite, 0)
 
 	clientHelloCanonical := canonicalPacketHandshake13(t, clientHello)
 	helloRetryRequestCanonical := canonicalPacketHandshake13(t, helloRetryRequest)
 
-	require.NoError(t, appendOutboundHandshakeFlight13(transcript, true, cipherSuite, []*dtlsflight.Packet{clientHello}))
-	require.NoError(t, appendOutboundHandshakeFlight13(
+	require.NoError(t, AppendOutboundHandshakeFlight(transcript, true, cipherSuite, []*dtlsflight.Packet{clientHello}))
+	require.NoError(t, AppendOutboundHandshakeFlight(
 		transcript, false, cipherSuite, []*dtlsflight.Packet{helloRetryRequest},
 	))
 
 	clientHelloHash := hashTranscript13(clientHelloCanonical)
 	messageHash := canonicalTranscriptHandshake13(handshake.TypeMessageHash, clientHelloHash)
 	expectedTranscript := append(append([]byte(nil), messageHash...), helloRetryRequestCanonical...)
-	assert.Equal(t, expectedTranscript, transcript.transcript)
+	assert.Equal(t, expectedTranscript, transcript.Bytes())
 
 	sum, err := transcript.sum()
 	require.NoError(t, err)
 	assert.Equal(t, hashTranscript13(messageHash, helloRetryRequestCanonical), sum)
-	require.Len(t, transcript.order, 2)
-	assert.Equal(t, handshake.TypeClientHello, transcript.order[0].typ)
-	assert.Equal(t, handshake.TypeServerHello, transcript.order[1].typ)
+	require.Len(t, transcript.messageOrder(), 2)
+	assert.Equal(t, handshake.TypeClientHello, transcript.messageOrder()[0].Type)
+	assert.Equal(t, handshake.TypeServerHello, transcript.messageOrder()[1].Type)
 
-	before := append([]byte(nil), transcript.transcript...)
-	require.NoError(t, appendOutboundHandshakeFlight13(
+	before := append([]byte(nil), transcript.Bytes()...)
+	require.NoError(t, AppendOutboundHandshakeFlight(
 		transcript, false, cipherSuite, []*dtlsflight.Packet{helloRetryRequest},
 	))
-	assert.Equal(t, before, transcript.transcript)
-	assert.Len(t, transcript.order, 2)
+	assert.Equal(t, before, transcript.Bytes())
+	assert.Len(t, transcript.messageOrder(), 2)
 }
 
 func transcriptTestClientHelloPacket13(sessionID []byte, seq uint16) *dtlsflight.Packet {
@@ -380,15 +432,17 @@ func transcriptTestClientHelloPacket13(sessionID []byte, seq uint16) *dtlsflight
 				Message: &handshake.MessageClientHello{
 					Version:            protocol.Version1_2,
 					SessionID:          sessionID,
-					CipherSuiteIDs:     []uint16{uint16(TLS_AES_128_GCM_SHA256)},
-					CompressionMethods: defaultCompressionMethods(),
+					CipherSuiteIDs:     []uint16{uint16(ciphersuite.TLS_AES_128_GCM_SHA256)},
+					CompressionMethods: dtlsflight.DefaultCompressionMethods(),
 				},
 			},
 		},
 	}
 }
 
-func transcriptTestHelloRetryRequestPacket13(tb testing.TB, cipherSuite CipherSuite, seq uint16) *dtlsflight.Packet {
+func transcriptTestHelloRetryRequestPacket13(
+	tb testing.TB, cipherSuite dtlsconfig.CipherSuite, seq uint16,
+) *dtlsflight.Packet {
 	tb.Helper()
 
 	random := handshake.Random{}
@@ -406,7 +460,7 @@ func transcriptTestHelloRetryRequestPacket13(tb testing.TB, cipherSuite CipherSu
 					Version:           protocol.Version1_2,
 					Random:            random,
 					CipherSuiteID:     &cipherSuiteID,
-					CompressionMethod: defaultCompressionMethods()[0],
+					CompressionMethod: dtlsflight.DefaultCompressionMethods()[0],
 					Extensions: []extension.Extension{
 						&extension.SupportedVersions{
 							Versions:        []protocol.Version{protocol.Version1_3},
